requirements.txt
fastapi==0.115.6
SQLAlchemy==2.0.36
psycopg2==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
uvicorn==0.32.1
bcrypt==4.2.1
PyJWT==2.10.1
alembic==1.14.0
